refactor: move Toaster into toast notifications component.
